+++ src/components/meal-card/UneditableMealCard.tsx
@@ -1,7 +1,7 @@
 import { Card, ListGroup } from "react-bootstrap";
 import { Meal } from "../../model/Food";
 import { FoodItem } from "./FoodItem";
-import { MealCalorieServingPanel } from "./MealCalorieServingPanel";
+import MealCalorieServingPanel from "../../features/meal-card/MealCalorieServingPanel";
 
 export const UneditableMealCard = (props: { meal: Meal; }) => {
   const { meal } = props;
